requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==20.3
sniffio==1.3.1
